+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBDescriptor.java
@@ -237,9 +237,9 @@ public class IoTDBDescriptor {
         conf.setMemtableSizeThreshold(memTableSizeThreshold);
       }
 
-      conf.setAvgSeriesPointNumberThreshold(Float.parseFloat(properties
+      conf.setAvgSeriesPointNumberThreshold(Integer.parseInt(properties
           .getProperty("avg_series_point_number_threshold",
-              Float.toString(conf.getAvgSeriesPointNumberThreshold()))));
+              Integer.toString(conf.getAvgSeriesPointNumberThreshold()))));
 
       conf.setSyncEnable(Boolean
           .parseBoolean(properties.getProperty("is_sync_enable",
